Collapse internal/k8s/typed/ into internal/k8s/
Move typed CRD types, client, scheme registration, and deepcopy from
the temporary internal/k8s/typed/ subpackage into internal/k8s/,
replacing the old dynamic client. Update all import paths across the
controller and main packages from typed.* to k8s.*.
