+++ internal/k8s/groupversion.go
@@ -1,4 +1,4 @@
-package typed
+package k8s
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
