+++ internal/k8s/types.go
@@ -1,4 +1,4 @@
-package typed
+package k8s
 
 import (
 	"fmt"
